Feature/exception handling

+++ BankSystem.Domain/Models/Currency.cs
@@ -4,6 +4,7 @@
     {
         public string Name { get; set; }
         public string Code { get; set; }
-        public decimal Price { get; set; }
+
+        public int Price { get; set; }
     }
 }
